Added a TodoItem class and Project factory
TodoItem is the class that is used to instantiate
todo objects. Every todo object is initialised
with a title, description, dueDate and priority.
The prototype for this class has methods for
changing the title, description, dueDate and
priority of a todo item.

The Project factory function is used to create
project objects for housing todo items. Projects
have an array property for storing and retrieving
todo items and methods for adding and removing
todo items, and chaning the project's title.